requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
APScheduler==3.11.2
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==3.1.2
Django==6.0.2
django-jazzmin==3.0.2
et_xmlfile==2.0.0
frozenlist==1.8.0
gunicorn==25.1.0
idna==3.11
multidict==6.7.1
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzlocal==5.3.1
urllib3==2.6.3
whitenoise==6.11.0
yarl==1.22.0
